+++ src/pari.cpp
@@ -3,7 +3,105 @@
 #include <fstream>
 #include <sstream>
 
+#include "expression_util.hpp"
 #include "log.hpp"
+
+void convertInitialTermsToIf(Formula& formula) {
+  auto it = formula.entries.begin();
+  while (it != formula.entries.end()) {
+    auto left = it->first;
+    Expression general(Expression::Type::FUNCTION, left.name,
+                       {Expression(Expression::Type::PARAMETER, "n")});
+    auto general_it = formula.entries.find(general);
+    if (left.type == Expression::Type::FUNCTION && left.children.size() == 1 &&
+        left.children.front()->type == Expression::Type::CONSTANT &&
+        general_it != formula.entries.end()) {
+      general_it->second =
+          Expression(Expression::Type::IF, "",
+                     {*left.children.front(), it->second, general_it->second});
+      it = formula.entries.erase(it);
+    } else {
+      it++;
+    }
+  }
+}
+
+void convertFracToPari(Expression& frac) {
+  std::string func = "floor";
+  if (ExpressionUtil::canBeNegative(*frac.children.at(0)) ||
+      ExpressionUtil::canBeNegative(*frac.children.at(1))) {
+    func = "truncate";
+  }
+  Expression wrapper(Expression::Type::FUNCTION, func, {frac});
+  frac = wrapper;
+}
+
+bool convertExprToPari(Expression& expr) {
+  // convert bottom-up!
+  for (auto& c : expr.children) {
+    if (!convertExprToPari(*c)) {
+      return false;
+    }
+  }
+  if (expr.type == Expression::Type::FRACTION) {
+    convertFracToPari(expr);
+  } else if (expr.type == Expression::Type::POWER) {
+    if (ExpressionUtil::canBeNegative(*expr.children.at(1))) {
+      Expression wrapper(Expression::Type::FUNCTION, "truncate", {expr});
+      expr = wrapper;
+    }
+  } else if (expr.type == Expression::Type::MODULUS) {
+    auto c1 = *expr.children.at(0);
+    auto c2 = *expr.children.at(1);
+    if (ExpressionUtil::canBeNegative(c1) ||
+        ExpressionUtil::canBeNegative(c2)) {
+      Expression wrapper(Expression::Type::DIFFERENCE);
+      wrapper.newChild(c1);
+      wrapper.newChild(Expression::Type::PRODUCT);
+      Expression frac(Expression::Type::FRACTION, "", {c1, c2});
+      convertFracToPari(frac);
+      wrapper.children[1]->newChild(c2);
+      wrapper.children[1]->newChild(frac);
+      expr = wrapper;
+    }
+  } else if (expr.type == Expression::Type::FUNCTION &&
+             expr.name == "binomial") {
+    // TODO: check feedback from PARI team to avoid this limitation
+    if (ExpressionUtil::canBeNegative(*expr.children.at(1))) {
+      return false;
+    }
+  }
+  return true;
+}
+
+bool Pari::convertToPari(Formula& f) {
+  for (auto& entry : f.entries) {
+    if (!convertExprToPari(entry.second)) {
+      return false;
+    }
+  }
+  convertInitialTermsToIf(f);
+  return true;
+}
+
+std::string Pari::toString(const Formula& f) {
+  std::string result;
+  bool first = true;
+  for (auto it = f.entries.rbegin(); it != f.entries.rend(); it++) {
+    if (!first) {
+      result += "; ";
+    }
+    if (f.entries.size() > 1) {
+      result += "(";
+    }
+    result += it->first.toString() + " = " + it->second.toString();
+    if (f.entries.size() > 1) {
+      result += ")";
+    }
+    first = false;
+  }
+  return result;
+}
 
 Sequence Pari::eval(const Formula& f, int64_t start, int64_t end) {
   const std::string gpPath("pari-loda.gp");
@@ -12,7 +110,7 @@ Sequence Pari::eval(const Formula& f, int64_t start, int64_t end) {
   if (!gp) {
     throw std::runtime_error("error generating gp file");
   }
-  gp << f.toString(true) << std::endl;
+  gp << toString(f) << std::endl;
   gp << "for (n = " << start << ", " << end << ", print(a(n)))" << std::endl;
   gp << "quit" << std::endl;
   gp.close();